@@ -64,9 +64,9 @@
     "proj4": "^2.4.3",
     "proj4leaflet": "^1.0.1",
     "prop-types": "^15.6.0",
-    "react": "^16.4.1",
+    "react": "^16.5.2",
     "react-datepicker": "^1.4.1",
-    "react-dom": "^16.4.1",
+    "react-dom": "^16.5.2",
     "react-dual-listbox": "^1.3.2",
     "react-element-to-jsx-string": "^13.2.0",
     "react-foundation": "^0.9.5",
